[hello-remote] split hello backend into robot-base and robot-camera backends
